chore: Working on figuring out layout
